+++ RELEASE_NOTES.md
@@ -18,6 +18,7 @@
 * `truncate` and `floor` on floats can be inlined given enough type information, if the resulting quotient is a fixnum.
 * `truncate` et al. on integers with unity divisor is optimized out.
 * Simple `gethash` calls are slightly cheaper.
+* Access to bit arrays and sub-byte arrays (e.g. `(unsigned-byte 4)`) can now be inlined.
 
 # Version 3.0.1 (LLVM15-20, 22) 2026-06-24